Revert "top nav idea"
